+++ cudax/test/stf/error_checks/write_frozen.cu
@@ -55,10 +55,10 @@ int main()
 #endif // !_CCCL_COMPILER_MSVC
 
   stream_ctx ctx;
-  const size_t N = 16;
+  const int N = 16;
   int X[N];
 
-  for (size_t i = 0; i < N; i++)
+  for (int i = 0; i < N; i++)
   {
     X[i] = i;
   }
